+++ immutable-arrays/transformations-to-standard-collections/src/main/kotlin/com/danrusu/pods4k/immutableArrays/TransformationsToSet.kt
@@ -1,0 +1,57 @@
+// Auto-generated file. DO NOT EDIT!
+package com.danrusu.pods4k.immutableArrays
+
+import kotlin.Boolean
+import kotlin.Byte
+import kotlin.Char
+import kotlin.Double
+import kotlin.Float
+import kotlin.Int
+import kotlin.Long
+import kotlin.Short
+import kotlin.collections.Set
+
+/**
+ * See [Array.toSet]
+ */
+public fun <T> ImmutableArray<T>.toSet(): Set<T> = asList().toSet()
+
+/**
+ * See [Array.toSet]
+ */
+public fun ImmutableBooleanArray.toSet(): Set<Boolean> = asList().toSet()
+
+/**
+ * See [Array.toSet]
+ */
+public fun ImmutableByteArray.toSet(): Set<Byte> = asList().toSet()
+
+/**
+ * See [Array.toSet]
+ */
+public fun ImmutableCharArray.toSet(): Set<Char> = asList().toSet()
+
+/**
+ * See [Array.toSet]
+ */
+public fun ImmutableShortArray.toSet(): Set<Short> = asList().toSet()
+
+/**
+ * See [Array.toSet]
+ */
+public fun ImmutableIntArray.toSet(): Set<Int> = asList().toSet()
+
+/**
+ * See [Array.toSet]
+ */
+public fun ImmutableLongArray.toSet(): Set<Long> = asList().toSet()
+
+/**
+ * See [Array.toSet]
+ */
+public fun ImmutableFloatArray.toSet(): Set<Float> = asList().toSet()
+
+/**
+ * See [Array.toSet]
+ */
+public fun ImmutableDoubleArray.toSet(): Set<Double> = asList().toSet()